requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
apscheduler==3.6.3
    # via python-telegram-bot
asgiref==3.5.2
    # via django
beautifulsoup4==4.11.1
    # via -r requirements.in
cachetools==4.2.2
    # via python-telegram-bot
certifi==2022.9.24
    # via
    #   python-telegram-bot
    #   requests
charset-normalizer==2.1.1
    # via requests
django==4.1.1
    # via -r requirements.in
huey==2.4.3
    # via -r requirements.in
idna==3.4
    # via requests
python-dotenv==0.21.0
    # via -r requirements.in
python-telegram-bot==13.14
    # via -r requirements.in
pytz==2022.2.1
    # via
    #   apscheduler
    #   python-telegram-bot
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
requests==2.28.1
    # via -r requirements.in
six==1.16.0
    # via apscheduler
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.3
    # via django
tornado==6.1
    # via python-telegram-bot
tzdata==2022.4
    # via pytz-deprecation-shim
tzlocal==4.2
    # via apscheduler
urllib3==1.26.12
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
